requirements.txt
asgiref==3.8.1
Django==5.1.7
django-bootstrap5==25.1
django-filter==25.1
djangorestframework==3.15.2
Markdown==3.7
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
